group writer tools
