+++ client/src/document/index.scss
@@ -1,4 +1,4 @@
-.loading-document-spy {
+.loading-toolbar {
   position: absolute;
   font-size: 70%;
   top: 0;
