--- client/src/document/spy.css
@@ -1,11 +1,0 @@
-#document-spy {
-  position: fixed;
-  font-size: 70%;
-  top: 0;
-  right: 0;
-  z-index: 2;
-  background-color: rgb(210, 210, 210);
-  opacity: 0.95;
-  padding: 3px;
-  border-radius: 0 0 3px 3px;
-}
