+++ client/src/document/writer-toolbar/edit-buttons.scss
@@ -1,4 +1,4 @@
-.edit-this-page {
+.edit-buttons {
   text-align: center;
   p.editor-opening-error {
     font-weight: bold;
