+++ client/src/document/writer-toolbar/flaws.scss
