+++ client/src/document/writer-toolbar/spy.css
@@ -1,0 +1,5 @@
+#document-spy {
+  font-size: 70%;
+  padding: 3px;
+  border-radius: 0 0 3px 3px;
+}
