+++ client/src/document/writer-toolbar/spy.tsx
